Update: Se eliminaron campos innecesarios en los recursos y modelos, se cambiaron tipos de datos a decimal en varias migraciones, y se ajustaron relaciones en el modelo DeudaCuota. Además, se renombró el campo 'nombre_completo' a 'nombre' en el controlador Inquilino.

+++ database/migrations/2024_08_16_144724_create_pagos_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->unsignedBigInteger('id_documento');
             $table->integer('numero_pago');
             $table->integer('serie');
-            $table->integer('total_pago');
+            $table->decimal('total_pago', 10, 2);
             $table->dateTime('fecha_registro')->useCurrent();
             
             // Relaciones
